pyproject.toml
[project]
name = "universal-agents"
version = "0.1.0"
description = "Self-evolving multi-agent framework for Claude Code"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "pydantic>=2.10",
    "rich>=13.9",
    "psutil>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/uagents"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
